# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ELLA - AI Personal Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 4 */
  password: string;
}

export interface AuthResponse {
  userId: number;
  username: string;
  token: string;
}

export interface JournalEntry {
  id: number;
  userId: number;
  content: string;
  /** @nullable */
  mood?: string | null;
  date: string;
  createdAt: string;
  updatedAt: string;
}

export interface JournalEntryInput {
  userId: number;
  content: string;
  mood?: string;
  date: string;
}

export interface JournalEntryUpdate {
  content?: string;
  mood?: string;
}

export interface Habit {
  id: number;
  userId: number;
  name: string;
  icon: string;
  color: string;
  createdAt: string;
}

export interface HabitInput {
  userId: number;
  name: string;
  icon: string;
  color: string;
}

export interface HabitLog {
  id: number;
  habitId: number;
  userId: number;
  date: string;
}

export interface HabitLogInput {
  userId: number;
  date: string;
}

export interface Goal {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: string;
  progress: number;
  /** @nullable */
  targetDate?: string | null;
  createdAt: string;
}

export interface GoalInput {
  userId: number;
  title: string;
  description?: string;
  category: string;
  targetDate?: string;
}

export interface GoalUpdate {
  title?: string;
  description?: string;
  category?: string;
  progress?: number;
  targetDate?: string;
}

export interface MoodLog {
  id: number;
  userId: number;
  mood: string;
  /** @nullable */
  note?: string | null;
  date: string;
  createdAt: string;
}

export interface MoodLogInput {
  userId: number;
  mood: string;
  note?: string;
  date: string;
}

export interface CalendarEvent {
  id: number;
  userId: number;
  title: string;
  date: string;
  type: string;
  createdAt: string;
}

export interface CalendarEventInput {
  userId: number;
  title: string;
  date: string;
  type: string;
}

export interface MeditationInput {
  userId: number;
  theme: string;
  duration: number;
  mood?: string;
}

export interface MeditationPreset {
  id: string;
  title: string;
  description: string;
  duration: number;
  theme: string;
  icon: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  userId: number;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
  userId: number;
}

export interface OpenaiMessageInput {
  content: string;
  userId: number;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  userId: number;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListJournalEntriesParams = {
userId: number;
};

export type ListHabitsParams = {
userId: number;
};

export type ListHabitLogsParams = {
userId: number;
};

export type ListGoalsParams = {
userId: number;
};

export type ListMoodLogsParams = {
userId: number;
};

export type ListCalendarEventsParams = {
userId: number;
};

export type ListOpenaiConversationsParams = {
userId: number;
};