# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ELLA - AI Personal Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List journal entries for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journal/{id}:
    get:
      operationId: getJournalEntry
      tags: [journal]
      summary: Get a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List habits for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /habits/{id}/log:
    post:
      operationId: logHabit
      tags: [habits]
      summary: Log a habit completion for a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitLogInput"
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitLog"
        "409":
          description: Already logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /habits/{id}/log/{date}:
    delete:
      operationId: unlogHabit
      tags: [habits]
      summary: Remove a habit log for a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /habits/logs:
    get:
      operationId: listHabitLogs
      tags: [habits]
      summary: List habit logs for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of habit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitLog"
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /calendar:
    get:
      operationId: listCalendarEvents
      tags: [calendar]
      summary: List calendar events for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createCalendarEvent
      tags: [calendar]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /calendar/{id}:
    delete:
      operationId: deleteCalendarEvent
      tags: [calendar]
      summary: Delete a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 4
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        token:
          type: string
      required:
        - userId
        - username
        - token
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        mood:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - content
        - date
        - createdAt
        - updatedAt
    JournalEntryInput:
      type: object
      properties:
        userId:
          type: integer
        content:
          type: string
        mood:
          type: string
        date:
          type: string
      required:
        - userId
        - content
        - date
    JournalEntryUpdate:
      type: object
      properties:
        content:
          type: string
        mood:
          type: string
    Habit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - icon
        - color
        - createdAt
    HabitInput:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
      required:
        - userId
        - name
        - icon
        - color
    HabitLog:
      type: object
      properties:
        id:
          type: integer
        habitId:
          type: integer
        userId:
          type: integer
        date:
          type: string
      required:
        - id
        - habitId
        - userId
        - date
    HabitLogInput:
      type: object
      properties:
        userId:
          type: integer
        date:
          type: string
      required:
        - userId
        - date
    Goal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        progress:
          type: integer
        targetDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - category
        - progress
        - createdAt
    GoalInput:
      type: object
      properties:
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        targetDate:
          type: string
      required:
        - userId
        - title
        - category
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        progress:
          type: integer
        targetDate:
          type: string
...
    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        date:
          type: string
        type:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - date
        - type
        - createdAt
    CalendarEventInput:
      type: object
      properties:
        userId:
          type: integer
        title:
          type: string
        date:
          type: string
        type:
          type: string
      required:
        - userId
        - title
        - date
        - type
...