# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ELLA - AI Personal Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  userId: number;
  title: string;
  description?: string;
  category: string;
  targetDate?: string;
}