# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * ELLA - AI Personal Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface Habit {
  id: number;
  userId: number;
  name: string;
  icon: string;
  color: string;
  createdAt: Date;
}